+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 0.1.35 (30-08-2021)
+- Fixed issue that caused invalid OBJ files to be produced for M2 collision geometry.
 - Added `skin.subMeshes.enabled` property to M2 metadata, indicating which submeshes were exported.
 - Added `groups.enabled` property to WMO metadata, indicating which WMO groups were exported.