+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Added `skin.subMeshes.enabled` property to M2 metadata, indicating which submeshes were exported.
 - Added `groups.enabled` property to WMO metadata, indicating which WMO groups were exported.
 - Added `textures.fileNameInternal` property to M2 metadata, mapping fileDataID to a listfile entry.
-- Added `textures.fileNameExternal` property to M2 metadata, pointing to relative texture export location.
+- Added `textures.fileNameExternal` property to M2 metadata, pointing to texture export location (relative to OBJ).
 - Added `textures.mtlName` property to M2 metadata, linking to an entry in the MTL file.
 
 0.1.34 (24-08-2021)